packages/phantom-visuals/pyproject.toml
# packages/phantom-visuals/pyproject.toml

[tool.poetry]
name = "phantom-visuals-v2"
version = "0.1.0"
description = "A project for phantom visuals"
authors = ["Pablo Aguirre <main@pabloagn.com>"]
readme = "README.md"
license = "LICENSE"
repository = "https://github.com/pabloagn/phantom/phantom-visuals-v2"
homepage = "https://github.com/pabloagn/phantom/phantom-visuals-v2"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
packages = [{include = "phantom_visuals_v2", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pillow = "^10.0.0"
numpy = "^1.24.0"
scikit-image = "^0.24.0"
opencv-python = "^4.8.0"
click = "^8.1.0"
rich = "^13.4.0"
pydantic = "^2.0.0"
colorama = "^0.4.6"
torch = "^2.5.0"
torchvision = "^0.20.0"
noise = "^1.2.0"
scipy = "^1.10.1"
matplotlib = "^3.7.1"
PyYAML = "^6.0"
mediapipe = "^0.10.0"
insightface = "^0.7.3"
onnxruntime-gpu = {version = "^1.15.0", optional = true}
kornia = "^0.6.9"
einops = "^0.6.1"
timm = "^0.9.2"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
tqdm = "^4.65.0"
numba = "^0.57.0"
cupy-cuda11x = {version = "^13.0.0,<14.0.0", optional = true}
pygfx = "^0.1.15"
trimesh = "^3.21.6"
pyrender = "^0.1.45"
# openmesh = ">=1.2.1"

[tool.poetry.group.dev.dependencies]
typer = "^0.9.0"
pytest = "^7.3.1"
mypy = "^1.3.0"
ruff = "^0.1.0"

# Entry points configuration
[tool.poetry.scripts]
phantom-visuals = "phantom_visuals_v2.cli:cli"
phantom-batch = "phantom_visuals_v2.cli:batch"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "C4",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "ANN", # flake8-annotations
    "N",   # pep8-naming
    "D",   # pydocstyle
    "UP",  # pyupgrade
    "TID", # flake8-tidy-imports
    "RUF", # Ruff-specific rules
    "PYI", # flake8-pyi
    "PT",  # flake8-pytest-style
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "ERA", # eradicate
    "ARG", # flake8-unused-arguments
    "PL",  # pylint
    "TRY", # tryceratops
]
ignore = [
    "D203",  # One blank line required before class docstring
    "D213",  # Multi-line docstring summary should start at the second line
]

[tool.ruff.per-file-ignores]
"tests/**/*.py" = [
    "D",     # No docstrings needed in tests
    "ANN",   # Type annotations not needed in tests
]
"__init__.py" = [
    "F401",  # Imported but unused
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.pycodestyle]
max-line-length = 88

[tool.ruff.mccabe]
max-complexity = 10

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
xvs_output = "native"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
